pyproject.toml
[project]
name = "binance_syncer"
version = "3.0.0"
description = "A high-performance synchronization tool for downloading and managing Binance historical data (klines, trades, etc.) asynchronously. It supports both local and S3 storage with intelligent data optimization (monthly vs daily preference)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "boto3>=1.39.8",
    "bs4>=0.0.2",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "requests>=2.32.4",
    "certifi>=2025.7.14",
    "rich>=14.0.0",
    "utilities-toolkit>=0.2.0",
    "duckdb>=1.4.3",
    "click>=8.1.8",
]

[project.scripts]
binance-syncer = "binance_syncer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/binance_syncer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.31.0",
]
